Cleanup exports. Rename Button index.
