+++ packages/components/package.json
@@ -4,6 +4,7 @@
   "license": "MIT",
   "private": true,
   "main": "build/index.js",
+  "sideEffects": false,
   "scripts": {
     "build": "NODE_ENV=production webpack --progress --config webpack.config.js",
     "watch": "NODE_ENV=development webpack --progress --watch --config webpack.config.js",
