--- packages/components/src/Utils/index.js
@@ -1,4 +1,0 @@
-import colors from './colors';
-import typography from './typography';
-
-export { colors, typography };
