+++ packages/components/src/index.js
@@ -1,5 +1,3 @@
-import { colors, typography } from './Utils';
-
-import AdminUIButton from './Components/Button';
-
-export { colors, typography, AdminUIButton };
+export { default as AdminUIBUtton } from './Components/Button/Button';
+export { default as colors } from './Utils/colors';
+export { default as typography } from './Utils/typography';
